Move pie module into donut subfolder
